Move coordinator to a new file
