+++ custom_components/sagemcom_fast/__init__.py
@@ -34,7 +34,7 @@ from .const import (
     LOGGER,
     PLATFORMS,
 )
-from .device_tracker import SagemcomDataUpdateCoordinator
+from .coordinator import SagemcomDataUpdateCoordinator
 
 SERVICE_REBOOT = "reboot"
