+++ custom_components/sagemcom_fast/device_tracker.py
@@ -1,25 +1,16 @@
 """Support for device tracking of client router."""
 from __future__ import annotations
 
-from datetime import timedelta
-import logging
-
-import async_timeout
 from homeassistant.components.device_tracker import SourceType
 from homeassistant.components.device_tracker.config_entry import ScannerEntity
-from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import StateType
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-    UpdateFailed,
-)
-from sagemcom_api.client import SagemcomClient
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from sagemcom_api.models import Device
 
 from .const import DOMAIN
+from .coordinator import SagemcomDataUpdateCoordinator
 
 
 async def async_setup_entry(hass, config_entry, async_add_entities):
@@ -33,53 +24,12 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
     )
 
 
-class SagemcomDataUpdateCoordinator(DataUpdateCoordinator):
-    """Class to manage fetching Sagemcom data."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        logger: logging.Logger,
-        *,
-        name: str,
-        client: SagemcomClient,
-        update_interval: timedelta | None = None,
-    ):
-        """Initialize update coordinator."""
-        super().__init__(
-            hass,
-            logger,
-            name=name,
-            update_interval=update_interval,
-        )
-        self.data = {}
-        self.hosts: dict[str, Device] = {}
-        self._client = client
-
-    async def _async_update_data(self) -> dict[str, Device]:
-        """Update hosts data."""
-        try:
-            async with async_timeout.timeout(10):
-                try:
-                    await self._client.login()
-                    hosts = await self._client.get_hosts(only_active=True)
-                finally:
-                    await self._client.logout()
-                """Mark all device as non-active."""
-                for idx, host in self.hosts.items():
-                    host.active = False
-                    self.hosts[idx] = host
-                for host in hosts:
-                    self.hosts[host.id] = host
-                return self.hosts
-        except Exception as exception:
-            raise UpdateFailed(f"Error communicating with API: {exception}")
-
-
-class SagemcomScannerEntity(ScannerEntity, RestoreEntity, CoordinatorEntity):
+class SagemcomScannerEntity(
+    ScannerEntity, RestoreEntity, CoordinatorEntity[SagemcomDataUpdateCoordinator]
+):
     """Sagemcom router scanner entity."""
 
-    def __init__(self, coordinator, idx, parent) -> None:
+    def __init__(self, coordinator: SagemcomDataUpdateCoordinator, idx, parent) -> None:
         """Initialize the device."""
         super().__init__(coordinator)
         self._idx = idx
